Remove explicit indication of whether a variable is in the state or not.
It turns out it always makes sense for the caller to be lazy about this, as the
environment always knows whether the variable being declared is part of the
global state or not. This simplification is possible because global state
elements can only be declared while we are in a 'bare' context (the upper most
scope).