separate build targets into a library and an executable
This commit moves towards a "double library" design that I'm aiming for. The
idea is to try and structure the code generator *and* the model as traditional
library-ised Linux programs. No attempt has been made to move towards the latter
yet; the present commit concerns only the former.

Currently generated files in the library aren't dealt with properly.
Specifically, they get generated in the output directory and we just hackily
expose them to the binary. They should really be generated at a path that CMake
is aware of needs to be exported. Might require some investigation into how to
achieve this.