librumur: fix: remove accidentally shadowing field
Well that was a rude surprise. VarDecl lookups in the symbol table failed
mysteriously and debugging indicated they had no name. Some digging revealed we
were accidentally shadowing Decl::name.

It is pretty surprising to me that -Wall -Wextra does not enable -Wshadow, which
would have caught this. Unfortunately we can't turn it on right now because it
sprays warnings about the constructors that shadow class members with their
arguments. Sigh. I guess I'll have to refactor this before turning it on.