librumur: enable referencing of quantified variables within forall
This works, but it makes me slightly nervous that the Var of a quantified
variable is only sustained by shared_ptrs of ExprIDs related to where it is
referenced. I.e. Once we traverse out of parsing the forall expression, the
enclosing scope is closed and the Var is destroyed if it hasn't been referenced.
On the one hand, this seems fine and might even help us later in detecting
useless quantification. On the other hand, the lifetime of this Var object is
quite slippery.