rewrite back into C++17
It turns out we can leverage some pretty powerful syntax and type resolution
features from C++17 to reduce the complexity of code generation. It's not
immediately clear to me whether this will result in degraded performance of
final code, but I think it's safer to bet on the C++ compiler's ability to
optimize code over my own. In the short term, this definitely simplifies our
task. I think a reasonable approach is to push forward in C++17 and, once we
have a working first version, consider whether porting to C11 would gain us
anything.