try a different strategy for emitting references
My thinking is that we can emit a single type that models both a reference and a
value. Which is in use will depend on the constructor called, which should
always be static. As a bonus, I think this should give us functions/procedures
for free without having to tweak this scheme.