use more indirection for constructing state references
It turns out there's a case inside guards (and possibly other contexts) where
the state reference we have is const and, for this reason, we want any
references to state variables to be const. We achieve this by calling an
overloaded static function that allows us to return a qualified constructed
object.