push validation into AST constructors
In service of validating AST nodes on construction, this commit does a number of
things:

  1. Remove indexing in Model constructors. This was unnecessary as indexing is
     already done during parsing.
  2. Implement operator== on Nodes.
  3. Refactor is_boolean to lean on TypeExpr::resolve rather than a direct
     comparison against the boolean type. This fixes problems wherein variables
     whose type was a TypeExprID referring to boolean were not considered
     booleans.
  4. Remove Node::validate and inline their functionality into respective
     constructors.

FIXME: There is some functionality in this commit that does not yet take
advantage of TypeExpr::resolve. For example, comparable() does not consider that
one of the expressions' types might be a TypeExprID that needs further
resolution.