support for iteration on range types
This supports basic iteration from MIN to MAX in the context of a for loop.
C++11 range-based for loops also work. We should further extend this to boolean
and user-defined enum types.

The motivation here is to be able to use for loops and similar iteration for
their equivalents in Murphi. In particular, this should allow us to emit forall
and exists statements as std::all_of and std::any_of. Aside from conciseness and
compiler friendliness, this also helps us avoid polluting the namespace.