separate Rule into its own class, SimpleRule
This leaves Rule as an abstract class which makes some logic more consistent.
For example, in the Rule (now SimpleRule) operator== we no longer need to
consider that the thing we're looking at might be a StartState. This commit also
fixes some minor issues. For example, Invariant inherited from Node and so could
not be stored in the Model's collection of rules.