allow Numbers to coerce to RangeBases
It's not clear to me whether this is a good idea or not. On the one hand, we
want this to happen in certain scenarios. Currently, when indexing arrays. On
the other hand, for numeric literals for the source to silently coerce to any
range type might be problematic.

My current feeling is that this shouldn't cause any problems. I *believe* any
numeric literal will only coerce to a range type in a scenario where we have
already validated that it *should* coerce to that type. I think this needs to be
proven out empirically though.