preliminary support for array element references in-state
When you reference an element of an array ("X[Y]") you want either an lvalue or
an rvalue. The prior skeleton subset operator ("operator[]") was pre-supposing
that these concepts would map naturally onto the C++ equivalents. That is, if
you are trying to write to an element of the array, you want a reference (&) and
if you are merely trying to read an element of the array, you want a value.

It turns out this is actually not accurate. When reading *or writing* we want a
value. The value contains -- within the object itself -- a reference to a
container to which it plans to write. The reference semantics of C++ have no
place in this domain and were actually a bit of a mirage. This discrepancy will
probably need more fixups.