separate generated type bases into 'reference' and 'value'
For each type, we had two possible representations: a local value and a
reference into the state bit vector. For example, ranges would become RangeBase
using either one set of its class members or the other. Instead of this
overloaded representation, we now have two separate classes (in this example
RangeReference and RangeValue) which both inherit from a common base (Range).

The previous unified representation was largely motivated by what I thought
would be a requirement when we implement procedures/functions. However, I no
longer think this is necessary and inheritance provides a neater way to phrase
this arrangement.