Array::operator[] return's ELEMENT_T::reference_type instead of ELEMENT_T&
The crucial enabling change in this commit was making ArrayValues BitBlocks.
With this in place, we no longer need to stick to returning an ELEMENT_T& from
Array::operator[]. We can now universally return a ELEMENT_T::reference_type
and, for local arrays, simple return a reference into the local.