remove STATE_T template parameter from Boolean
This was the logical thing to do to match what we did with Range. Unfortunately
we need to retain a dummy template parameter to Boolean to force it to be a
template. The reason for this, as explained in the comment, is that we don't
want it to be instantiated until its children (BooleanReference and
BooleanValue) are defined as we refer to them within Boolean. Messy.