rename TypeExpr::size to width
This is part of a fix to a problem I just noticed where we conflate the width of
a type and the number of elements it contains when calculating the size of an
array type. Instead what we should really be doing is aligning with what's
happening in generated code; i.e. having distinct concepts of "width" and
"count." This is a step towards that.

Related to Github #2: "array access out of bounds"