fix: missing bool assignment for Boolean
When you start doing boolean operations (!, ||, &&, ...) on a Boolean it decays
to an unboxed bool. It makes sense therefore to allow assignment to a Boolean
from an unboxed bool. We may decide later down the line that we want to retain
boxing during boolean operations, but I doubt it as there's no equivalent of
arithmetic overflow or similar that we want to catch here.