remove unused QueueLink
This was previously introduced when the multithreaded queue was built on
atomics. It's now built on std::queue, so we have no need for the manual
linked-list.

Though it was not the purpose of this commit, this change has a noticeable
improvement on performance as well. Our current work horse, long-running2, is
down from 51 seconds single-threaded to 46 seconds. With 4 threads, it drops
from 85 seconds to 68 seconds.

Related to Github #6 "Multithreaded optimisation"