expose closed set implementation for use
This doesn't quite work yet, because we allocate the closed set on the stack and
all its members (i.e. the set itself) are inline. The result is that we overflow
the stack and immediately segfault when using any reasonably large set size. We
need to slightly refactor the code to stick the set in BSS.