treat all references as being into the state
Instead of previously relying on the abstract interface BitBlock as a container
that references would point into, we now assume all references are into the
state specifically. This makes things much cleaner and in fact removes the need
for BitBlock altogether.

The previous motivation for BitBlock was to support condensed arrays (array
locals that were internally dense bit vectors). I've come around to thinking
this is not necessary because we can just represent such local arrays naively as
C++ structs.

Closes Github #10 "why is ArrayValue a BitBlock?"