fix segfault on reference to local
The way the code generation logic was phrased, references to local variables of
enum type would accidentally be considered members of the enum. With this
refactoring we now correctly treat these as references to locals. However,
things are still not completely correct because we never actually define the
local we are referencing.