fix: read of uninitialised pointer in Quantifier::clone
When a quantifier was duplicated -- which happened as part of Model::generate --
the clone logic was assuming Quantifier::step was either null or a valid
pointer. In the case where a quantifier had resulted from a variable ranging
over a previously declared type by using the type's ID, Quantifier::step was not
being initialised at all. This initially exhibited as an illegal instruction
execution during model generation of the German protocol [0]. Further debugging
revealed that the cause was this errant pointer. The present commit fixes this
issue by now initialising the pointer during construction.

  [0]: https://bitbucket.org/jderick/preach/src/ec841f695d323c853e28bfad89ec7512fb95f002/test/german.m?at=default&fileviewer=file-view-default