fix: duplicate ruleset quantifiers when appending to flattened rules
When a ruleset produces a set of flattened rules during Ruleset::flatten, the
resulting rules are intended to be independent copies. However, the quantifiers
of the newly generated rules were accidentally shallow copies of the quantifiers
from the containing ruleset. As a result, deletion of the flattened rules would
leave some dangerous dangling pointers in the original ruleset.

This was a latent issue in the current generator as we don't do anything with
the model after generation and never explicitly delete it. However, if a ruleset
had more than one containing rule this would have resulted in a double free. In
future it would have caused more severe errors as we go on to use models in more
advanced ways.