fix incorrect assertion
A typo in this assertion made it fire incorrectly for certain models. Because we
turn assertions into optimisation hints in a non-debug build, this actually
turned out to be the root cause of Github #35. Interesting that GCC managed to
do so much with this incorrect hint as to turn the program into an infinite
loop.

Fixes Github #35 "No progress in optimised checker"