periodically check if another thread has found an error
This should be the penultimate step to preparing to go fully multithreaded. The
one remaining wart is that we can take an exit during the start state phase in
which no secondary threads will be active. The master thread will try to join
all the secondary threads, none of which will have started yet. To be handled in
an upcoming commit.