only score the element type of arrays for interdependence in scalarset pivots
Prior to this commit, we were scoring an array type used in a symmetry reduction
pivot by both its index type *and* element type. This is fine, but it means that
pivot components were never scoring below a 1 for interdependence. Still this is
OK, until we start considering how this interacts with component comparison. A
perfect pivot with three components would score a 3, while an imperfect pivot
with a single component could score a two. This would lead us to incorrectly
conclude the perfect pivot was worse than the imperfect one.

To remedy this we now score such array types based solely on their element type.
Using this scheme, a perfect pivot always scores 0.

In this commit we also now cache interdependence calculations. This wasn't for
performance reasons, but just led to more elegant code.

Github: related to #12 "symmetry reduction"