fix: unify TypeExpr::width()
This fixes numerous errors and inconsistencies, including:

  * simple types with only a single value ("undefined") were being allocated one
    bit of storage unnecessarily;
  * the width required for a Range type was not accounting for "undefined";
  * the count of a Scalarset type was assuming the bound itself was a valid
    value;
  * records did not cope well with zero-width members;
  * arrays were allocating width for "undefined" of the index type;
  * arrays did not cope well with zero-width indices or elements.

By unifying the way in which we calculate a type's width we get some
consistency. This is still not a complete fix as the calculations will now
overflow for large types. Hopefully this will no longer be an issue if we move
to GMP.

After these changes, the size of german.m's state drops from 149 bits to 86
bits.

FIXME: there is also still an outstanding bug with this calculation. The size of
the german.m model is computed as 86 bits, even though my manual calculations
suggest that it should be 80 bits.