fix: canonicalise current state being worked on
During exploration we were accidentally applying canonicalisation to the
*previous* state rather than the current one. This caused major chaos because
the previous state was in the state set and thus we ended up changing its value
post-insertion. We should reduce the chance of this kind of error happening in
future by returning a const pointer from dequeue.

Github: fixes #51 "symmetry reduction broken"