per-thread queues
We now have a state queue per-thread and the checking logic directs which queue
to enqueue into. On dequeue, we look at the last queue we looked at or walk
forward from there to find a new state. Note that we still do not start any new
threads; this is still just preparation work.