fix: ensure correct set count in multithreaded execution
Under a multithreaded checker, it is possible for one thread to see set
occupancy exceed the expansion threshold and initiate expansion while another
thread(s) are in the insertion loop. The resulting race can look something like
this:

  1. Thread A starts insertion loop
  2. Thread B starts insertion loop
  3. Thread A finishes insertion
  4. Thread A starts a new insertion, but sees high occupancy
  5. Thread A allocates a new set (and does new->count = old->count)
  6. Thread B finishes insertion (old->count++)

The result of this, at quiescence, is that the set count is now one less than it
should be. To resolve this problem, we now only set the count of the new set at
the completion of migration and when there is only one remaining reference to
the old set.