disable sandbox by default
Sandboxing gets somewhat complicated depending on the platform. On Linux,
sandboxing is done via Seccomp which is difficult to use in a portable way.
Across Linux kernel versions and libc vendors, the syscalls used to implement a
given libc function will vary, and yet the Seccomp interface is to tell the
kernel which syscalls you are going to use. Rather than bending over backwards
to implement portable sandboxing, we turn sandboxing off by default and let
users who want this protection enable it manually. The reality is that few will
enable it, but the reality is also that *very* few will be running adversarial
models. Interestingly -- based on a macOS example -- sandbox has a significant
overhead as well.

Github: related to #50 "sandbox checker"