decouple data structure invariant checking from construction
The idea here is to allow invariants to be violated or at least not satisfiable
on construction of an object. The motivation for this is to do two-phase parsing
and symbol resolution. We would parse the input into an AST, then as a separate
pass resolve symbols to what they point to.

Github: related to #64 "separate symbol resolution from parsing"