add new more rigorous base traversal interfaces
The purpose of this circuitousness is to introduce a class that can serve as a
parent that *requires* the child to override all 'visit' methods. For some
traversals -- e.g. the AST dumper -- we want to override everything. More
precisely, we would like compilation to fail if we don't override everything.
This structure allows us to achieve this.