remove Lvalue abstract class
This class was misleading. Specifically, ExprID inherited from Lvalue, but
whether an ExprID could serve as an lvalue or not depended on its referent. That
is, whether something is an lvalue or not cannot be determined from the class
hierarchy alone. To resolve this, we just remove the Lvalue inheritance
indication altogether and implement a new is_lvalue method.