move output function to rumur binary
This is the start of a process of moving all the checker output code out of the
library and into the binary. I realised while implementing some of the more
involved parts of checker generation that this code is not at all reusable and
doesn't make sense in the library. When the checker is complete it may make
sense to generalise some pieces for insertion back into the library.