support passing nullptr to Range constructor
It turns out we need to support a range type that is unbounded. The use case for
this is to describe the type of a quantifier variable whose bounds ('from' and
'to') are not constant. It seems like we should probably extend this to using
this type (an unbounded range) to describe numeric literals as well.

The implementation is a bit of a hack, in that we hard code the limits to assume
the eventual underlying type is int64_t. We should really support 'min' and
'max' being nullptr in Range.