anticipate VarDecls having a null type
I'm still undecided whether this is a good idea or not. I initially introduced
this to support quantifiers where the loop variable has non-constant bounds, but
I'm thinking in the long term it may be better to just give these cases proper
(non-null) types. In other words, we may end up backing out this commit.