support for quantified variables with non-constant bounds
Surprisingly there are models that use for loops quantified with a non-constant
lower and upper limit. To support this we need to drop the idea of building a
type for the loop variable, because it does not necessarily have a known width.
Support in this commit is a little hacky and it would be nice not to hard code
so many aspects of value_t.