fix: check invariants post-canonicalisation
By checking invariants before canonicalisation (the previous logic), it was
possible to trigger an invariant violation whose enabling rule could not be
found. The sequence of events would look like:

  1. Take transition rule
  2. Check invariants and discover violation; call error()
  3. print_transition() tries all rules from the preceding state *with*
     canonicalisation
  4. The *canonicalised* end state of every rule fails to match the
     non-canonicalised state that triggered the invariant violation.

We are not out of the woods yet... This issue was found while troubleshooting
Github #73. While debugging this I realised it is also possible for a mid-rule
assertion to fail, throwing an error state that does not match the result of any
rule transition. Worse, the print_transition() logic will actually *re-trigger*
this error, probably causing infinite recursion.

Github: Related to #73 "n_peterson causes bus error"