fix: rewrite print_counterexample to avoid recursion
In the case of a lengthy counterexample trace, it was possible to cause a stack
overflow when calling print_counterexample. Rather than try to think of a more
clever solution to this, it seems straightforward to just anticipate that (a) a
counterexample trace can be of arbitrary length and (b) printing a
counterexample trace is not a performance critical operation. With this in mind,
we rephrase print_counterexample to avoid recursion and heap-allocate some
memory to handle printing the trace. This now makes it possible to check
CMurphi's ex/mux/n_peterson.m, though we currently get the wrong result.

Github: closes #73 "n_peterson causes bus error"