run test suite directly with verboseness in Travis CI
Travis seems to consider no output after 10 minutes a stuck job and terminates a
build. This has now become a problem that the test suite can exceed 10 minutes
and the default ctest mode is to suppress all output. To work around this, we
force the test suite to run in verbose mode dumping a lot of output to the
terminal. Not ideal, but meh.