treat Property::Category as an int in Bison
It seems Bison tries to read from the rule result variable ($$) before we've had
a chance to initialise it, triggering undefined behaviour. I can't figure out
how to get around this with an enum, so I've taken the crude escape hatch and
used a bare number.

This issue was identified by UBSan.