switch to an algorithm for exhaustive canonicalisation
This replaces the previous canonicalisation algorithm (which was buggy) with an
algorithm that uses exhaustive permutations based on Heap's algorithm. This is
extremely expensive -- in terms of runtime -- but at least gives a baseline
(hopefully) correct symmetry reduction algorithm. On german.m with 5 NODE and 5
DATA this drops the number of states explored to 131112. This also ends up
being a runtime win as we finish in 464s.

Note that this removes the concept of "pivots" in symmetry-reduction.cc. This
was getting in the way more than adding clarity, and was obscuring some of the
above mentioned bugs. I have some ideas of introducing a more coherent and
useful idea of pivots in a future commit.

Github: related to #12 "symmetry reduction"