abstract canonicalisation by another level
This gives us a top level state_canonicalise function in the checker that just
serves as a branching point to another canonicalisation function that does the
real work. The purpose of this is to pave the way for introducing a faster
heuristic canonicalisation function as an option.

Github: related to #12 "symmetry reduction"