fix CMurphi expected results in test suite
It was a mistake to mark n_peterson as expecting to fail, due to an issue with
symmetry reduction. On CMurphi, this model passes. There is still a minor issue
in that ex/tmp/scalarset.m is marked as if it should succeed when really it
should fail.

Github: related to #79 "deadlock detection on looping rules"