implement CMurphi's semantics for deadlock
In addition to the previous two modes of deadlock detection (off and on), we
implement a "stuttering" mode. This considers any state that only has enabled
transitions that lead back to itself also deadlocked. This new mode, which
matches CMurphi's semantics, becomes the default.

Note that in this commit we had to adjust several test cases, which assumed the
previous "stuck" semantics as well as the expected outcome of two CMurphi tests
which was wrong. This commit also reveals the symmetry reduction problem more as
the CMurphi models mcslock1.m and mcslock2.m (both in ex/mux and ex/sym)
incorrectly fail with a deadlock detected in this mode, while they pass without
using symmetry reduction.

Github: closes #79 "deadlock detection on looping rules"
Github: related to #12 "symmetry reduction"