implement a handle_read_raw option without using __int128
This rewrites one of the two unconditional uses of __int128 (the other being
handle_write_raw), to support 32-bit platforms. In the process of implementing
this, I discovered that reading two 64-bit values and then shifting and ORing
them together seems noticeably more expensive than reading a single 128-bit
value on x86-64. To try to get the best of both worlds, I've rewritten this to
use __int128 when possible and more performant.