generate state variable handles in state_print
The thinking behind this is that we can use these handles to print VarDecls in
state_print, instead of emitting the handles inline. The advantage of this is
that it enables generate_print to be used outside of the context of state_print
to satisfy a put statement.

Github: related to #84 "implement puts statements"