new generate_print parameter for whether to support diffing
This commit adds a new boolean parameter to the generate_print function
indicating whether we need to anticipate diffing the previous state's value and
suppressing the print if it is not different. The point of all of this is to
move this function closer to what we need for implementing the put statement.