fix: validate model before reindexing
This is one of those "how did it ever work?" kind of things. Reindexing calls
VarDecl::width(), which in turn calls a number of functions that assume the
model has already been validated. By pure luck we seem to manage to avoid
triggering an assertion failure or bad reference. Interestingly this was found
while debugging another orthogonal issue, when my initial fix for that issue
failed.

Github: related to #91 "inverted ranges are accepted"