fix Github release v2018.12.08
Github automatically bundles up the source of a version-tagged commit on its
releases tab. Unfortunately the provided zip and tarball don't compile because
they do not contain any Git information or a version file. We manually add a
version file in this commit (which we'll remove in the following commit) to fix
this issue.