flip the version logic in make-version
This commit changes the priority of how we select a Rumur version to:
  1. Version tag
  2. Commit hash
  3. Last known release

The purpose of this is to avoid the awkward juggling we had to do with
v2018.12.08. By prioritising the Git version tag, all we need to do for a
release is (1) bump the version number and (2) tag the commit and Github's
automated release process should actually work properly for us.