fix: support boolean constants
Previously it was possible to define a boolean constant, but then this could not
easily be used. Rumur considered all ConstDecls without an explicit type (i.e.
non-enum member ConstDecls) to have the type of the unbounded range, null. As a
result, boolean constants were not usable in boolean expressions because Rumur
believed they were not booleans.

Github: closes #98 "boolean consts are not usable"