handle non-lvalues passed in as var parameters
Consider the following example:

  procedure foo(var x: boolean); begin
    ...
  end;
  ...
  foo(true);

It's not clear to me exactly what the semantics of this are intended to be.
Perhaps this is intended to be rejected as invalid. However, it seems clear what
an author who wrote this intended. As with C++ semantics, construct a temporary
that gets discarded after the function call.

In this commit we implement these semantics manually. We generate a temporary
within the verifier to permit expressions like these to work as expected.

Github: closes #100 "passing boolean literal as var parameter causes uncaught
exception"