implement isundefined
It would have been nice to just implement this as a built-in function, rather
than making the lexer and parser aware of it. However, we can't do this because
it needs to take a generic lvalue (of any simple type). Perhaps an argument
could be made for adding support for generics, but for now the path of least
resistance was teaching the lexer that isundefined is now a keyword.

Github: closes #96 "implement isundefined"